attribute_matching_service/requirements.txt
fastapi
uvicorn
pandas
numpy
requests
google-cloud-storage
google-cloud-aiplatform
pyyaml
rapidfuzz
fuzzywuzzy
